Refactored lib file names to new standard
