+++ cipher-lib.js
