+++ gatekeeper-lib.js
@@ -2,7 +2,7 @@ import { json } from '@helia/json';
 import { base58btc } from 'multiformats/bases/base58';
 import canonicalize from 'canonicalize';
 import { createHelia } from 'helia';
-import * as cipher from './cipher.js';
+import * as cipher from './cipher-lib.js';
 import config from './config.js';
 
 const validVersions = [1];
