+++ gatekeeper.test.js
@@ -1,5 +1,5 @@
 import mockFs from 'mock-fs';
-import * as cipher from './cipher.js';
+import * as cipher from './cipher-lib.js';
 import * as gatekeeper from './gatekeeper-lib.js';
 import * as db_json from './db-json.js';
